requirements.txt
Django==5.1.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==23.0.0
python-dotenv==0.16.0